adds react router and basic routes (index 404)
