+++ app/client/components/page-not-found.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default class PageNotFound extends React.Component {
+  render() {
+    return (<p>nothing here</p>)
+  }
+}
