+++ app/client/main.js
@@ -1,5 +1,21 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+import { browserHistory, Router, Route, IndexRoute, Link } from 'react-router'
 import App from './App.js'
+import PageNotFound from 'components/page-not-found'
+import Dashboard from 'components/dashboard'
 
-ReactDOM.render(<div><App /></div>, document.getElementById('root'));
+var Root = React.createClass({
+  render: function() {
+    return (
+      <Router history={browserHistory}>
+        <Route path="/" component={App}>
+          <IndexRoute component={Dashboard} />
+        </Route>
+        <Route path="*" component={PageNotFound} />
+      </Router>
+    )
+  }
+})
+
+ReactDOM.render(<Root />, document.getElementById('root'));
